Restructure evidence artifact contracts
