+++ packages/ai/nina/workspace/preflight.test.ts
@@ -1,19 +1,15 @@
 // @vitest-environment node
 
 import { MATH_CAPABILITY } from "@repo/ai/nina/capability/spec";
-import { PEDAGOGY_MOVE_SUMMARY_MAX_LENGTH } from "@repo/ai/nina/pedagogy/schema";
 import {
   decodeEvidenceWorkspace,
   EVIDENCE_CONTRIBUTION_ARTIFACT_BYTES,
   EVIDENCE_CONTRIBUTION_ARTIFACT_LIMIT,
-  EVIDENCE_CONTRIBUTION_MODEL_SUMMARY_MAX_LENGTH,
-  EVIDENCE_CONTRIBUTION_PEDAGOGY_MOVE_LIMIT,
   EVIDENCE_WORKSPACE_ARTIFACT_BYTES,
   EVIDENCE_WORKSPACE_ARTIFACT_LIMIT,
-  EVIDENCE_WORKSPACE_TURN_ID_MAX_LENGTH,
   EvidenceWorkspaceDecodeError,
 } from "@repo/ai/nina/workspace/schema";
-import { MAX_COORDINATE_ARTIFACT_BYTES } from "@repo/math/schema/artifact";
+import { MAX_COORDINATE_ARTIFACT_BYTES } from "@repo/math/schema/artifact/safety";
 import { Cause, Effect, Exit, Option } from "effect";
 import { describe, expect, it } from "vitest";
 
@@ -195,145 +191,6 @@ describe("EvidenceWorkspace invariants", () => {
 
     expectDecodeFailure(failure, "Invalid evidence workspace contract.");
   });
-
-  it("bounds model-visible contribution and pedagogy text", async () => {
-    const longSummary = await decodeFailure(
-      workspace([
-        contribution({
-          modelSummary: "x".repeat(
-            EVIDENCE_CONTRIBUTION_MODEL_SUMMARY_MAX_LENGTH + 1
-          ),
-        }),
-      ])
-    );
-
-    expectDecodeFailure(longSummary, "Invalid evidence workspace contract.");
-
-    const tooManyMoves = await decodeFailure(
-      workspace([
-        contribution({
-          pedagogyMoves: Array.from(
-            { length: EVIDENCE_CONTRIBUTION_PEDAGOGY_MOVE_LIMIT + 1 },
-            () => pedagogyMove("cas://math/evaluate")
-          ),
-        }),
-      ])
-    );
-
-    expectDecodeFailure(tooManyMoves, "Invalid evidence workspace contract.");
-
-    const longMoveSummary = await decodeFailure(
-      workspace([
-        contribution({
-          pedagogyMoves: [
-            {
-              ...pedagogyMove("cas://math/evaluate"),
-              summary: "x".repeat(PEDAGOGY_MOVE_SUMMARY_MAX_LENGTH + 1),
-            },
-          ],
-        }),
-      ])
-    );
-
-    expectDecodeFailure(
-      longMoveSummary,
-      "Invalid evidence workspace contract."
-    );
-  });
-
-  it("bounds workspace turn identifiers", async () => {
-    const blank = await decodeFailure(workspace([], { turnId: "   " }));
-    expectDecodeFailure(blank, "Invalid evidence workspace contract.");
-
-    const oversized = await decodeFailure(
-      workspace([], {
-        turnId: "t".repeat(EVIDENCE_WORKSPACE_TURN_ID_MAX_LENGTH + 1),
-      })
-    );
-    expectDecodeFailure(oversized, "Invalid evidence workspace contract.");
-  });
-
-  it("blocks artifacts and pedagogy moves on unavailable evidence", async () => {
-    const failedArtifact = await decodeFailure(
-      workspace([
-        contribution({
-          artifacts: [artifact("artifact-from-failed-evidence")],
-          evidenceStatus: "failed",
-        }),
-      ])
-    );
-
-    expectDecodeFailure(
-      failedArtifact,
-      "Contribution math cannot attach artifacts when evidence status is failed."
-    );
-
-    const deniedPedagogy = await decodeFailure(
-      workspace([
-        contribution({
-          evidenceStatus: "denied",
-          pedagogyMoves: [pedagogyMove("cas://math/evaluate")],
-        }),
-      ])
-    );
-
-    expectDecodeFailure(
-      deniedPedagogy,
-      "Contribution math cannot attach pedagogy moves when evidence status is denied."
-    );
-
-    const decoded = await Effect.runPromise(
-      decodeEvidenceWorkspace(
-        workspace([contribution({ evidenceStatus: "failed" })])
-      )
-    );
-    expect(decoded.contributions[0]?.evidence.status).toBe("failed");
-  });
-
-  it("validates pedagogy evidence refs against contribution refs and artifacts", async () => {
-    const ungrounded = await decodeFailure(
-      workspace([
-        contribution({
-          pedagogyMoves: [pedagogyMove("missing-ref")],
-        }),
-      ])
-    );
-
-    expectDecodeFailure(
-      ungrounded,
-      "Pedagogy move verification-note references unknown evidence missing-ref."
-    );
-
-    const decoded = await Effect.runPromise(
-      decodeEvidenceWorkspace(
-        workspace([
-          contribution({
-            artifacts: [artifact("artifact-1")],
-            pedagogyMoves: [pedagogyMove("artifact-1")],
-          }),
-        ])
-      )
-    );
-
-    expect(decoded.contributions[0]?.pedagogyMoves?.[0]?.evidenceRefs).toEqual([
-      "artifact-1",
-    ]);
-  });
-
-  it("allows pedagogy refs grounded only by artifact ids", async () => {
-    const decoded = await Effect.runPromise(
-      decodeEvidenceWorkspace(
-        workspace([
-          contributionWithoutEvidenceRefs({
-            artifacts: [artifact("artifact-only")],
-            pedagogyMoves: [pedagogyMove("artifact-only")],
-          }),
-        ])
-      )
-    );
-
-    expect(decoded.contributions[0]?.evidence.refs).toBeUndefined();
-  });
 });
 
 function workspace(
@@ -347,42 +204,17 @@ function workspace(
   };
 }
 
-function contribution(
-  input: {
-    artifacts?: readonly unknown[];
-    evidenceStatus?: string;
-    modelSummary?: string;
-    pedagogyMoves?: readonly unknown[];
-  } = {}
-) {
+function contribution(input: { artifacts?: readonly unknown[] } = {}) {
   return {
     artifacts: input.artifacts,
     capability: MATH_CAPABILITY,
     evidence: {
       capability: MATH_CAPABILITY,
       refs: ["cas://math/evaluate"],
-      status: input.evidenceStatus ?? "available",
-      summary: "CAS verified the coordinate relation.",
-    },
-    modelSummary: input.modelSummary ?? "CAS verified the coordinate relation.",
-    pedagogyMoves: input.pedagogyMoves,
-  };
-}
-
-function contributionWithoutEvidenceRefs(input: {
-  artifacts: readonly unknown[];
-  pedagogyMoves: readonly unknown[];
-}) {
-  return {
-    artifacts: input.artifacts,
-    capability: MATH_CAPABILITY,
-    evidence: {
-      capability: MATH_CAPABILITY,
       status: "available",
       summary: "CAS verified the coordinate relation.",
     },
     modelSummary: "CAS verified the coordinate relation.",
-    pedagogyMoves: input.pedagogyMoves,
   };
 }
 
@@ -423,14 +255,6 @@ function artifact(id: string, description?: string) {
     },
     proofAnchors: ["cas://coordinate/artifact"],
     title: "Coordinate artifact",
-  };
-}
-
-function pedagogyMove(evidenceRef: string) {
-  return {
-    evidenceRefs: [evidenceRef],
-    kind: "verification-note",
-    summary: "Anchor the explanation in verified evidence.",
   };
 }
