+++ packages/math/schema/artifact/schema.ts
@@ -1,30 +1,21 @@
-import { decodeMathAst, ExactScalar } from "@repo/math/schema/ast";
+import { findLearningArtifactInvariantIssue } from "@repo/math/schema/artifact/invariant";
 import {
-  type CoordinatePrimitive,
+  findRawArtifactSizeIssue,
+  MAX_COORDINATE_ARTIFACT_PRIMITIVES,
+  MAX_COORDINATE_ARTIFACT_PROOF_ANCHOR_LENGTH,
+  MAX_COORDINATE_ARTIFACT_PROOF_ANCHORS,
+  MAX_LEARNING_ARTIFACT_ID_LENGTH,
+} from "@repo/math/schema/artifact/safety";
+import { decodeMathAst, ExactScalar } from "@repo/math/schema/ast/schema";
+import {
   CoordinatePrimitiveSchema,
   RenderSamplingPolicy,
   readCoordinatePrimitiveMathAsts,
-} from "@repo/math/schema/coordinate-primitives";
-import { readSortableExactScalar } from "@repo/math/schema/coordinate-scalars";
-import { findCoordinatePrimitiveIssue } from "@repo/math/schema/coordinate-validation";
+} from "@repo/math/schema/coordinate/primitive";
+import { findCoordinatePrimitiveIssue } from "@repo/math/schema/coordinate/validation";
 import { Effect, Schema } from "effect";
 
 export const COORDINATE_SYSTEM_ARTIFACT_KIND = "coordinate-system-3d";
-
-/** Maximum serialized size accepted for one coordinate learning artifact. */
-export const MAX_COORDINATE_ARTIFACT_BYTES = 750_000;
-
-/** Maximum deterministic primitive count accepted for one coordinate artifact. */
-export const MAX_COORDINATE_ARTIFACT_PRIMITIVES = 64;
-
-/** Maximum length accepted for one learning artifact id. */
-export const MAX_LEARNING_ARTIFACT_ID_LENGTH = 180;
-
-/** Maximum proof anchors accepted on one coordinate artifact. */
-export const MAX_COORDINATE_ARTIFACT_PROOF_ANCHORS = 16;
-
-/** Maximum length accepted for one proof anchor reference. */
-export const MAX_COORDINATE_ARTIFACT_PROOF_ANCHOR_LENGTH = 180;
 
 const ArtifactIdSchema = Schema.NonEmptyString.pipe(
   Schema.pattern(/\S/),
@@ -92,11 +83,17 @@ export class LearningArtifactDecodeError extends Schema.TaggedError<LearningArti
   }
 ) {}
 
-/** Decodes a learning artifact and verifies symbolic math references. */
+/**
+ * Decodes a learning artifact and verifies symbolic math references.
+ */
 export const decodeLearningArtifact = Effect.fn(
   "math.artifact.decodeLearningArtifact"
 )(function* (input: unknown) {
-  const rawSizeIssue = yield* readRawArtifactSizeIssue(input);
+  const rawSizeIssue = yield* findRawArtifactSizeIssue(input).pipe(
+    Effect.mapError(
+      (error) => new LearningArtifactDecodeError({ message: error.message })
+    )
+  );
   if (rawSizeIssue) {
     return yield* Effect.fail(
       new LearningArtifactDecodeError({ message: rawSizeIssue })
@@ -114,19 +111,10 @@ export const decodeLearningArtifact = Effect.fn(
     )
   );
 
-  const duplicateId = findDuplicatePrimitiveId(artifact.payload.primitives);
-  if (duplicateId) {
+  const artifactIssue = findLearningArtifactInvariantIssue(artifact);
+  if (artifactIssue) {
     return yield* Effect.fail(
-      new LearningArtifactDecodeError({
-        message: `Duplicate coordinate primitive id: ${duplicateId}.`,
-      })
-    );
-  }
-
-  const axisIssue = findAxisRangeIssue(artifact);
-  if (axisIssue) {
-    return yield* Effect.fail(
-      new LearningArtifactDecodeError({ message: axisIssue })
+      new LearningArtifactDecodeError({ message: artifactIssue })
     );
   }
 
@@ -151,76 +139,3 @@ export const decodeLearningArtifact = Effect.fn(
 
   return artifact;
 });
-
-function findDuplicatePrimitiveId(primitives: readonly CoordinatePrimitive[]) {
-  const ids = new Set<string>();
-
-  for (const primitive of primitives) {
-    if (ids.has(primitive.id)) {
-      return primitive.id;
-    }
-    ids.add(primitive.id);
-  }
-}
-
-function readRawArtifactSizeIssue(input: unknown) {
-  return Effect.try({
-    catch: () =>
-      new LearningArtifactDecodeError({
-        message: "Invalid learning artifact contract.",
-      }),
-    try: () => JSON.stringify(input),
-  }).pipe(
-    Effect.flatMap((json) =>
-      json === undefined
-        ? Effect.fail(
-            new LearningArtifactDecodeError({
-              message: "Invalid learning artifact contract.",
-            })
-          )
-        : Effect.succeed(findJsonSizeIssue(json))
-    )
-  );
-}
-
-function findJsonSizeIssue(json: string) {
-  const sizeBytes = new TextEncoder().encode(json).byteLength;
-
-  if (sizeBytes > MAX_COORDINATE_ARTIFACT_BYTES) {
-    return `Coordinate artifact exceeds ${MAX_COORDINATE_ARTIFACT_BYTES} bytes.`;
-  }
-}
-
-function findAxisRangeIssue(artifact: LearningArtifact) {
-  const xIssue = findOneAxisRangeIssue("x", artifact.payload.axes.x);
-  if (xIssue) {
-    return xIssue;
-  }
-
-  const yIssue = findOneAxisRangeIssue("y", artifact.payload.axes.y);
-  if (yIssue) {
-    return yIssue;
-  }
-
-  return findOneAxisRangeIssue("z", artifact.payload.axes.z);
-}
-
-function findOneAxisRangeIssue(
-  axisName: "x" | "y" | "z",
-  range: readonly [ExactScalar, ExactScalar]
-) {
-  const min = readSortableExactScalar(range[0]);
-  const max = readSortableExactScalar(range[1]);
-
-  if (min === undefined) {
-    return `Coordinate artifact ${axisName}-axis range must use sortable numeric bounds.`;
-  }
-
-  if (max === undefined) {
-    return `Coordinate artifact ${axisName}-axis range must use sortable numeric bounds.`;
-  }
-
-  if (min >= max) {
-    return `Coordinate artifact ${axisName}-axis range must be increasing.`;
-  }
-}
