--- packages/math/schema/ast.test.ts
@@ -1,369 +1,0 @@
-import { decodeMathAst, MathAstDecodeError } from "@repo/math/schema/ast";
-import { Cause, Effect, Exit, Option } from "effect";
-import { describe, expect, it } from "vitest";
-
-describe("MathAst", () => {
-  it("decodes a flat symbolic expression graph", async () => {
-    const ast = await Effect.runPromise(
-      decodeMathAst({
-        canonical: "x + x",
-        latex: "x + x",
-        nodes: [
-          {
-            id: "x",
-            kind: "variable",
-            name: "x",
-          },
-          {
-            id: "sum",
-            kind: "binary",
-            left: "x",
-            operator: "add",
-            right: "x",
-          },
-        ],
-        root: "sum",
-      })
-    );
-
-    expect(ast.root).toBe("sum");
-    expect(ast.nodes).toHaveLength(2);
-  });
-
-  it("decodes unary expressions with valid operand references", async () => {
-    const ast = await Effect.runPromise(
-      decodeMathAst({
-        canonical: "-x",
-        latex: "-x",
-        nodes: [
-          {
-            id: "x",
-            kind: "variable",
-            name: "x",
-          },
-          {
-            id: "negate-x",
-            kind: "unary",
-            operand: "x",
-            operator: "negate",
-          },
-        ],
-        root: "negate-x",
-      })
-    );
-
-    expect(ast.root).toBe("negate-x");
-  });
-
-  it("rejects invalid schema shapes with a typed error", async () => {
-    const exit = await Effect.runPromiseExit(
-      decodeMathAst({
-        canonical: "1",
-        latex: "1",
-        nodes: [],
-        root: "one",
-      })
-    );
-
-    const failure = readExitFailure(exit);
-
-    expect(failure).toBeInstanceOf(MathAstDecodeError);
-    if (failure instanceof MathAstDecodeError) {
-      expect(failure.message).toBe("Invalid MathAst contract.");
-    }
-  });
-
-  it("rejects blank exact scalar expressions with a typed error", async () => {
-    const exit = await Effect.runPromiseExit(
-      decodeMathAst({
-        canonical: " ",
-        latex: " ",
-        nodes: [
-          {
-            id: "blank",
-            kind: "literal",
-            value: scalar("   "),
-          },
-        ],
-        root: "blank",
-      })
-    );
-
-    const failure = readExitFailure(exit);
-
-    expect(failure).toBeInstanceOf(MathAstDecodeError);
-    if (failure instanceof MathAstDecodeError) {
-      expect(failure.message).toBe("Invalid MathAst contract.");
-    }
-  });
-
-  it("rejects blank node ids with a typed error", async () => {
-    const exit = await Effect.runPromiseExit(
-      decodeMathAst({
-        canonical: "1",
-        latex: "1",
-        nodes: [
-          {
-            id: "   ",
-            kind: "literal",
-            value: scalar("1"),
-          },
-        ],
-        root: "   ",
-      })
-    );
-
-    const failure = readExitFailure(exit);
-
-    expect(failure).toBeInstanceOf(MathAstDecodeError);
-    if (failure instanceof MathAstDecodeError) {
-      expect(failure.message).toBe("Invalid MathAst contract.");
-    }
-  });
-
-  it("rejects nonnumeric literal values with a typed error", async () => {
-    const exit = await Effect.runPromiseExit(
-      decodeMathAst({
-        canonical: "left",
-        latex: "left",
-        nodes: [
-          {
-            id: "literal-left",
-            kind: "literal",
-            value: scalar("left"),
-          },
-        ],
-        root: "literal-left",
-      })
-    );
-
-    const failure = readExitFailure(exit);
-
-    expect(failure).toBeInstanceOf(MathAstDecodeError);
-    if (failure instanceof MathAstDecodeError) {
-      expect(failure.message).toBe(
-        "MathAst literal node literal-left must use a sortable numeric value."
-      );
-    }
-  });
-
-  it("rejects duplicate node ids with a typed error", async () => {
-    const exit = await Effect.runPromiseExit(
-      decodeMathAst({
-        canonical: "1",
-        latex: "1",
-        nodes: [
-          {
-            id: "one",
-            kind: "literal",
-            value: scalar("1"),
-          },
-          {
-            id: "one",
-            kind: "literal",
-            value: scalar("1"),
-          },
-        ],
-        root: "one",
-      })
-    );
-
-    const failure = readExitFailure(exit);
-
-    expect(failure).toBeInstanceOf(MathAstDecodeError);
-    if (failure instanceof MathAstDecodeError) {
-      expect(failure.message).toBe("Duplicate MathAst node id: one.");
-    }
-  });
-
-  it("rejects missing root nodes with a typed error", async () => {
-    const exit = await Effect.runPromiseExit(
-      decodeMathAst({
-        canonical: "1",
-        latex: "1",
-        nodes: [
-          {
-            id: "one",
-            kind: "literal",
-            value: scalar("1"),
-          },
-        ],
-        root: "missing",
-      })
-    );
-
-    const failure = readExitFailure(exit);
-
-    expect(failure).toBeInstanceOf(MathAstDecodeError);
-    if (failure instanceof MathAstDecodeError) {
-      expect(failure.message).toBe("MathAst root node was not found: missing.");
-    }
-  });
-
-  it("rejects cyclic graphs with a typed error", async () => {
-    const exit = await Effect.runPromiseExit(
-      decodeMathAst({
-        canonical: "-(-x)",
-        latex: "-(-x)",
-        nodes: [
-          {
-            id: "first",
-            kind: "unary",
-            operand: "second",
-            operator: "negate",
-          },
-          {
-            id: "second",
-            kind: "unary",
-            operand: "first",
-            operator: "negate",
-          },
-        ],
-        root: "first",
-      })
-    );
-
-    const failure = readExitFailure(exit);
-
-    expect(failure).toBeInstanceOf(MathAstDecodeError);
-    if (failure instanceof MathAstDecodeError) {
-      expect(failure.message).toBe(
-        "MathAst graph contains a cycle at node first."
-      );
-    }
-  });
-
-  it("rejects unreachable nodes with a typed error", async () => {
-    const exit = await Effect.runPromiseExit(
-      decodeMathAst({
-        canonical: "1",
-        latex: "1",
-        nodes: [
-          {
-            id: "one",
-            kind: "literal",
-            value: scalar("1"),
-          },
-          {
-            id: "x",
-            kind: "variable",
-            name: "x",
-          },
-        ],
-        root: "one",
-      })
-    );
-
-    const failure = readExitFailure(exit);
-
-    expect(failure).toBeInstanceOf(MathAstDecodeError);
-    if (failure instanceof MathAstDecodeError) {
-      expect(failure.message).toBe("MathAst node is unreachable from root: x.");
-    }
-  });
-
-  it("rejects missing unary operands with a typed error", async () => {
-    const exit = await Effect.runPromiseExit(
-      decodeMathAst({
-        canonical: "-x",
-        latex: "-x",
-        nodes: [
-          {
-            id: "negate-x",
-            kind: "unary",
-            operand: "x",
-            operator: "negate",
-          },
-        ],
-        root: "negate-x",
-      })
-    );
-
-    const failure = readExitFailure(exit);
-
-    expect(failure).toBeInstanceOf(MathAstDecodeError);
-    if (failure instanceof MathAstDecodeError) {
-      expect(failure.message).toBe(
-        "MathAst unary node negate-x references missing operand x."
-      );
-    }
-  });
-
-  it("rejects missing node references with a typed error", async () => {
-    const exit = await Effect.runPromiseExit(
-      decodeMathAst({
-        canonical: "x + 1",
-        latex: "x + 1",
-        nodes: [
-          {
-            id: "sum",
-            kind: "binary",
-            left: "x",
-            operator: "add",
-            right: "one",
-          },
-        ],
-        root: "sum",
-      })
-    );
-
-    const failure = readExitFailure(exit);
-
-    expect(failure).toBeInstanceOf(MathAstDecodeError);
-    if (failure instanceof MathAstDecodeError) {
-      expect(failure.message).toBe(
-        "MathAst binary node sum references missing left operand x."
-      );
-    }
-  });
-
-  it("rejects missing right operands with a typed error", async () => {
-    const exit = await Effect.runPromiseExit(
-      decodeMathAst({
-        canonical: "x + 1",
-        latex: "x + 1",
-        nodes: [
-          {
-            id: "x",
-            kind: "variable",
-            name: "x",
-          },
-          {
-            id: "sum",
-            kind: "binary",
-            left: "x",
-            operator: "add",
-            right: "one",
-          },
-        ],
-        root: "sum",
-      })
-    );
-
-    const failure = readExitFailure(exit);
-
-    expect(failure).toBeInstanceOf(MathAstDecodeError);
-    if (failure instanceof MathAstDecodeError) {
-      expect(failure.message).toBe(
-        "MathAst binary node sum references missing right operand one."
-      );
-    }
-  });
-});
-
-function scalar(expression: string) {
-  return {
-    expression,
-    latex: expression,
-  };
-}
-
-/** Extracts the typed Effect failure from an Exit for schema assertions. */
-function readExitFailure(exit: Exit.Exit<unknown, unknown>) {
-  if (Exit.isSuccess(exit)) {
-    return;
-  }
-
-  const failure = Cause.failureOption(exit.cause);
-  return Option.isSome(failure) ? failure.value : undefined;
-}
