+++ packages/math/schema/ast/pi.ts
@@ -1,13 +1,10 @@
-import { readExactNumericExpression } from "@repo/math/schema/coordinate-exact-numeric";
+import { readExactNumericExpression } from "@repo/math/schema/coordinate/numeric";
 
 const PI_TOKEN_PATTERN = /π|pi/gi;
 
-interface PiMultipleOperand {
-  piMultiple?: number;
-  value: number;
-}
-
-/** Reads a syntactic finite multiple of pi from the exact scalar grammar. */
+/**
+ * Reads a syntactic finite multiple of pi from the exact scalar grammar.
+ */
 export function readSyntacticPiMultiple(expression: string, value: number) {
   const compact = expression.trim().replaceAll(/\s+/g, "");
   const matches = compact.match(PI_TOKEN_PATTERN);
@@ -20,27 +17,37 @@ export function readSyntacticPiMultiple(expression: string, value: number) {
     : undefined;
 }
 
-/** Returns true only for exact integer multiples tracked syntactically. */
+/**
+ * Returns true only for exact integer multiples tracked syntactically.
+ */
 export function isSyntacticIntegerPiMultiple(multiple: number | undefined) {
   return multiple !== undefined && Number.isInteger(multiple);
 }
 
-/** Returns true only for exact half-integer multiples tracked syntactically. */
+/**
+ * Returns true only for exact half-integer multiples tracked syntactically.
+ */
 export function isSyntacticHalfIntegerPiMultiple(multiple: number | undefined) {
   return multiple !== undefined && Number.isInteger(multiple - 0.5);
 }
 
-/** Carries syntactic pi multiples through unary negation. */
+/**
+ * Carries syntactic pi multiples through unary negation.
+ */
 export function readNegatedPiMultiple(multiple: number | undefined) {
   return multiple === undefined ? undefined : -multiple;
 }
 
-/** Carries syntactic pi multiples through absolute value. */
+/**
+ * Carries syntactic pi multiples through absolute value.
+ */
 export function readAbsolutePiMultiple(multiple: number | undefined) {
   return multiple === undefined ? undefined : Math.abs(multiple);
 }
 
-/** Carries syntactic pi multiples through addition and subtraction. */
+/**
+ * Carries syntactic pi multiples through addition and subtraction.
+ */
 export function readCombinedPiMultiple(
   left: number | undefined,
   right: number | undefined,
@@ -53,10 +60,12 @@ export function readCombinedPiMultiple(
   return operator === "add" ? left + right : left - right;
 }
 
-/** Carries one syntactic pi factor through multiplication by a numeric scalar. */
+/**
+ * Carries one syntactic pi factor through multiplication by a numeric scalar.
+ */
 export function readProductPiMultiple(
-  left: PiMultipleOperand,
-  right: PiMultipleOperand
+  left: { readonly piMultiple?: number; readonly value: number },
+  right: { readonly piMultiple?: number; readonly value: number }
 ) {
   if (left.piMultiple !== undefined && right.piMultiple === undefined) {
     return left.piMultiple * right.value;
@@ -67,10 +76,12 @@ export function readProductPiMultiple(
   }
 }
 
-/** Carries a syntactic pi numerator through division by a numeric scalar. */
+/**
+ * Carries a syntactic pi numerator through division by a numeric scalar.
+ */
 export function readQuotientPiMultiple(
-  left: PiMultipleOperand,
-  right: PiMultipleOperand
+  left: { readonly piMultiple?: number; readonly value: number },
+  right: { readonly piMultiple?: number; readonly value: number }
 ) {
   return left.piMultiple === undefined || right.piMultiple !== undefined
     ? undefined
