+++ packages/math/schema/coordinate/numeric.ts
